requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-telegram-bot==20.7
google-generativeai==0.3.2
groq==0.4.1
httpx==0.25.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
firebase-admin==6.3.0
structlog==24.1.0
pytest==7.4.3
pytest-asyncio==0.21.1